movies_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
ipython==7.26.0
ipython-genutils==0.2.0
factory-boy==3.2.0
Faker==8.10.3
